backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
qdrant-client==1.7.0
sentence-transformers==2.2.2
ollama==0.1.7
pandas==2.1.4
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.2.0
python-multipart==0.0.6
aiofiles==23.2.1
numpy==1.26.2
torch>=2.0.0
transformers>=4.30.0
watchdog==3.0.0
schedule==1.2.0

# Security
PyJWT==2.8.0
bleach==6.1.0
slowapi==0.1.9

# Additional
rich==13.7.0
beautifulsoup4==4.12.2